設定画面の検索・削除処理エラー修正

+++ app/Http/Controllers/SettingController.php
@@ -8,31 +8,27 @@ class SettingController extends Controller
 {
 
     public function setting_index_get(Request $request){
-        $itemList = Item::all();
+        $itemList = Item::paginate(2);
         return view('setting.index', compact('itemList'));
     }
-    public function setting_index_delete(Request $request){
-        $deleteId = Item::find($request->deleteId)->delete();
-        $itemList = Item::all();
-        return view('setting.index', compact('itemList'));
-    }
-    public function setting_index_update(Request $request){
-        $deleteId = Item::find($request->deleteId)->delete();
-        $itemList = Item::all();
-        return view('setting.index', compact('itemList'));
-    }
-    public function setting_index_search(Request $request){
-        $searchItem = Item::query();
-        $searchNumber = $request->searchItemNumber;
-        $searchName = $request->searchItemName;
-        if(!empty($searchNumber)){
-            $searchItem->where('item_name','like','%'.$searchNumber.'%');
+    public function setting_index_post(Request $request){
+        if(isset($request->deleteId)){
+            $deleteId = Item::find($request->deleteId)->delete();
+            $itemList = Item::all();
+            return view('setting.index', compact('itemList'));
+        }else{
+            $searchItem = Item::query();
+            $searchNumber = $request->searchItemNumber;
+            $searchName = $request->searchItemName;
+            if(!empty($searchNumber)){
+                $searchItem->where('item_name','like','%'.$searchNumber.'%');
+            }
+            if(!empty($searchName)){
+                $searchItem->where('item_name','like','%'.$searchName.'%');
+            }
+            $searchData = $searchItem->get();
+            return view('setting.index', compact('searchData'));
         }
-        if(!empty($searchName)){
-            $searchItem->where('item_name','like','%'.$searchName.'%');
-        }
-        $searchData = $searchItem->get();
-        return view('setting.index', compact('searchData'));
     }
     public function setting_create_get(Request $request){
         return view('setting.create');
@@ -55,7 +51,6 @@ class SettingController extends Controller
         unset($getUpdateData['_token']);
         $updateData = Item::find($request->id);
         $updateData->fill($getUpdateData)->save();
-        $updateData = Item::find($request->updateId);
         return view('setting.details', compact('updateData'));
     }
 }
